redis_task/requirements.txt
blinker==1.8.2
click==8.1.7
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
pymssql==2.3.0
python-dotenv==1.0.1
redis==5.0.5
Werkzeug==3.0.3
